+++ src/common/services/recordGenerator/schemas/profiles/index.ts
@@ -1,3 +1,4 @@
+import { authorityRecordSchema } from './authority.schema';
 import { hubRecordSchema } from './hub.schema';
 import { instanceRecordSchema } from './instance.schema';
 import { workRecordSchema } from './work.schema';
@@ -6,4 +7,5 @@ export const profileRecordSchemas = {
   instance: instanceRecordSchema,
   work: workRecordSchema,
   hub: hubRecordSchema,
+  authority: authorityRecordSchema,
 };