+++ src/views/Search/config/index.ts
@@ -1,2 +1,3 @@
 export { createResourceActionsConfig, type ResourceActionsParams } from './resourcesActions.config';
 export { createHubActionsConfig, type HubActionsParams } from './hubsActions.config';
+export { createAuthorityActionsConfig, type AuthorityActionsParams } from './authorityActions.config';